bevy_reflect: Fix binary deserialization not working for unit structs (#6722) [skip ci]
Fixes #6713

Binary deserialization is failing for unit structs as well as structs with all ignored/skipped fields.

Add a check for the number of possible fields in a struct before deserializing. If empty, don't attempt to deserialize any fields (as there will be none).

Note: ~~This does not apply to enums as they do not properly handle skipped fields (see #6721).~~ Enums still do not properly handle skipped fields, but I decided to include the logic for it anyways to account for `#[reflect(ignore)]`'d fields in the meantime.

---

- Fix bug where deserializing unit structs would fail for non-self-describing formats